More on Functions - use "generic functions" instead of "generic"
E.g. sentence "Remember, _generic_ are all about relating two or more values with the same type!" is incorrect - consider an absolutely valid generic from the next section:
```
interface Box<Type> {
  contents: Type;
}
```
This is because "More on Functions" section includes information on "generic functions" and not about "generics". This PR would fix this kind of false statements